scripts/whisper-project/pyproject.toml
[project]
name = "whisper-subtitles"
version = "0.2.0"
description = "Generate subtitles from video files using OpenAI Whisper"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai-whisper>=20231117",
    "torch>=2.0.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "ruff>=0.1.0",
]

[project.scripts]
whisper-subtitles = "whisper_subtitles.cli:run_cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "W",    # pycodestyle warnings
    "UP",   # pyupgrade
]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
